Posture_Corrector_Remake
I am beginning this project willing to understand in depth how an accelerometer works, using an Arduino Uno and MPU5060 in this first setup. 
Reading this mockup was possible with a Superb Tech youtube video, where a library and its code were used to build mine. The code's only significant change was the delay time on the void loop section to have a quicker reading of the x, y, and z variables.